Added modifiers for setting backgrounds
Also added some extra code in list-based initializers because the collection view's `backgroundColor` doesn't initially do anything when using a list-based layout. That's because it uses the list configuration's `backgroundColor` on top of the collection view's `backgroundColor`. To get around this, the list configuration's `backgroundColor` will default to `.clear` (but can be overridden in the `listConfigurationHandler`), and internal background color (`collectionViewBackgroundColor`) will default to the list appearance's default background color.